Enum reqwless::client::HttpConnection

pub enum HttpConnection<'conn, C>
where
    C: Read + Write,
{
    Plain(C),
    PlainBuffered(BufferedWrite<'conn, C>),
    Tls(TlsConnection<'conn, C, Aes128GcmSha256>),
}

Represents a HTTP connection that may be encrypted or unencrypted.

Variants

Plain(C)

PlainBuffered(BufferedWrite<'conn, C>)

Tls(TlsConnection<'conn, C, Aes128GcmSha256>)

Implementations

impl<'conn, T> HttpConnection<'conn, T>
where T: Read + Write,

pub fn into_buffered<'buf>( self, tx_buf: &'buf mut [u8], ) -> HttpConnection<'buf, T>
where 'conn: 'buf,

Turn the request into a buffered request.

This is only relevant if no TLS is used, as embedded-tls buffers internally and we reuse its buffer for non-TLS connections.

pub async fn send<'req, 'buf, B: RequestBody>( &'conn mut self, request: Request<'req, B>, rx_buf: &'buf mut [u8], ) -> Result<Response<'conn, 'buf, HttpConnection<'conn, T>>, Error>

Send a request on an established connection.

The request is sent in its raw form without any base path from the resource. The response headers are stored in the provided rx_buf, which should be sized to contain at least the response headers.

The response is returned.