Struct reqwless::client::HttpResourceRequestBuilder

pub struct HttpResourceRequestBuilder<'req, 'conn, C, B>
where
    C: Read + Write,
    B: RequestBody,
{ /* private fields */ }

Implementations

impl<'req, 'conn, C, B> HttpResourceRequestBuilder<'req, 'conn, C, B>
where C: Read + Write, B: RequestBody,

pub async fn send<'buf>( self, rx_buf: &'buf mut [u8], ) -> Result<Response<'req, 'buf, HttpConnection<'conn, C>>, Error>

Send the request.

The base path of the resource is prepended to the request path. The response headers are stored in the provided rx_buf, which should be sized to contain at least the response headers.

The response is returned.